### 2023-08-15 #5761 @editorialbot
[PRE REVIEW]: GridFormat: header-only C++-library for grid file I/O
**Submitting author:** @dglaeser (<a href="http://orcid.org/0000-0001-9646-881X">Dennis Gläser</a>)
**Repository:** https://github.com/dglaeser/gridformat
**Branch with paper.md** (empty if default branch): feature/paper
**Version:** v0.1
