### 2023-08-15 #5761 @danielskatz
👋 @dglaeser - note that your paper is about twice as long as recommended for a JOSS paper. Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and consider if there is material in the paper that should be in the software documentation instead, and with that referred to in the paper. In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. You can also use the command `@editorialbot check repository` to see how long your paper is. editorialbot commands need to be the first entry in a new comment.
