+++ paper/paper.bib
@@ -147,6 +147,7 @@ isbn={978-3-0348-8534-8}
   title={{ParaView}: An end-user tool for large-data visualization},
   author={Ahrens, James and Geveci, Berk and Law, Charles},
   journal={The Visualization Handbook},
+  doi={10.1016/B978-012387582-2/50038-1},
   pages={},
   year={2005},
   month={01}
