### 2023-08-15 c5d1cd2
[paper] remove minimal example
