+++ paper/paper.md
@@ -131,133 +131,8 @@ files. `GridFormat` provides access to the data as specified by the file format,
 in all applications. For instance, to fully instantiate a simulator for parallel computations, information on the grid entities shared
 by different processes is usually required. Since these requirements are simulator-specific, any further processing has to be done
 manually by the user and for their data structures. The recommended way to deal with this issue is to add any information required for
-reinstantiation as data fields to the output. This way, it is readily available when reading the file.
-
-## Minimal Example
-
-Let us assume that we have some hand-written code that performs simulations on a two-dimensional, structured grid
-and yields a `std::vector<double>` storing the numerical solution. Due to this known trivial topology, the code does not
-define a specific data structure to represent grids. It only needs to know the number of cells per direction and the size of one cell.
-In the following code snippets, we demonstrate how one can use `GridFormat` to write such data into established file formats,
-in this case, the [VTI](https://examples.vtk.org/site/VTKFileFormats/#imagedata) file format. Stitched together, the subsequent snippets
-constitute a fully working C++ program.
-
-First, let us define a simple data structure to represent the grid:
-
-```cpp
-#include <array>
-
-struct MyGrid {
-    std::array<std::size_t, 2> cells;
-    std::array<double, 2> dx;
-};
-```
-
-We can specialize the traits required for the `ImageGrid` concept
-for the type `MyGrid` (see the code comments for more details).
-
-```cpp
-// Include the GridFormat library header (brings in the traits declarations)
-#include <gridformat/gridformat.hpp>
-
-namespace GridFormat::Traits {
-
-// Expose a range over grid cells. Here, we simply use the MDIndexRange provided
-// by GridFormat, which allows iterating over all index tuples within the given
-// dimensions (in our case the number of cells in each coordinate direction)
-// MDIndexRange yields objects of type MDIndex, and thus, GridFormat deduces
-// that MDIndex is the cell type of `MyGrid`.
-template<> struct Cells<MyGrid> {
-    static auto get(const MyGrid& grid) {
-        return GridFormat::MDIndexRange{{grid.cells[0], grid.cells[1]}};
-    };
-};
-
-// Expose a range over grid points
-template<> struct Points<MyGrid> {
-    static auto get(const MyGrid& grid) {
-        return GridFormat::MDIndexRange{{grid.cells[0]+1, grid.cells[1]+1}};
-    };
-};
-
-// Expose the number of cells of the image grid per coordinate direction
-template<> struct Extents<MyGrid> {
-    static auto get(const MyGrid& grid) {
-        return grid.cells;
-    }
-};
-
-// Expose the size of the cells per coordinate direction
-template<> struct Spacing<MyGrid> {
-    static auto get(const MyGrid& grid) {
-        return grid.dx;
-    }
-};
-
-// Expose the position of the grid origin
-template<> struct Origin<MyGrid> {
-    static auto get(const MyGrid& grid) {
-        // The origin for our grid's coordinate system is (0, 0)
-        return std::array{0.0, 0.0};
-    }
-};
-
-// For a given point or cell, expose its location (i.e. index tuple) within
-// the structured grid arrangement. Our point/cell types are the same, i.e.
-// GridFormat::MDIndex, since we used MDIndexRange in the Points/Cells traits.
-template<> struct Location<MyGrid, GridFormat::MDIndex> {
-    static auto get(const MyGrid& grid, const GridFormat::MDIndex& i) {
-        return std::array{i.get(0), i.get(1)};
-    }
-};
-
-}  // namespace GridFormat::Traits
-```
-
-With these traits defined, `MyGrid` can now be written out into file formats that are designed for image grids.
[…]
