### 2023-08-15 #5761 @dglaeser
@danielskatz, thanks for your comment. I fixed the missing DOI and removed the example code shown in the paper (I would move that into the examples in the repository). The word count seems to be 1467 now, I understand I should try to reduce it to 1000? I guess I can further shorten it by referencing things in the documentation.
