### 2023-08-15 #5761 @danielskatz
I don't think further reductions are needed - the paper now looks fine for review to me.
