### 2023-08-15 #5761 @danielskatz
@danasolav - September starts in just 2 weeks 🙂

If you could start then, I would be happy to assign it to you now with the understanding you wouldn't start immediately.
