### 2023-08-15 6d731ac
[examples] example2 -> dem
