+++ examples/README.md
@@ -28,7 +28,7 @@ be interested in.
 - [voxels](./voxels): register a data structure of voxels as `ImageGrid`
 - [dem](./dem): register a data structure representing a digital elevation model (dem) as `StructuredGrid`
 - [cgal](./cgal): using the predefined traits for [CGAL](https://www.cgal.org/).
-- [Example 4](./example4): using the predefined traits for [Dune::GridView](https://dune-project.org/).
+- [dune](./dune): using the predefined traits for [Dune::GridView](https://dune-project.org/).
 - [Example 5](./example5): using the predefined traits for [dolfinx](https://github.com/FEniCS/dolfinx).
 - [Example 6](./example6): using the predefined traits for [MFEM](https://mfem.org/).
 - [Example 7](./example7): reading/writing grid files in parallel computations using [MPI](https://de.wikipedia.org/wiki/Message_Passing_Interface).
