### 2023-08-15 7e6809b
[examples] example7 -> parallel
