+++ examples/README.md
@@ -31,4 +31,4 @@ be interested in.
 - [dune](./dune): using the predefined traits for [Dune::GridView](https://dune-project.org/).
 - [dolfinx](./dolfinx): using the predefined traits for [dolfinx](https://github.com/FEniCS/dolfinx).
 - [mfem](./mfem): using the predefined traits for [MFEM](https://mfem.org/).
-- [Example 7](./example7): reading/writing grid files in parallel computations using [MPI](https://de.wikipedia.org/wiki/Message_Passing_Interface).
+- [parallel](./parallel): reading/writing grid files in parallel computations using [MPI](https://de.wikipedia.org/wiki/Message_Passing_Interface).
