+++ examples/CMakeLists.txt
@@ -42,7 +42,7 @@ endif ()
 
 find_package(MPI)
 if (MPI_FOUND)
-    add_subdirectory(example7)
+    add_subdirectory(parallel)
 else ()
-    message(STATUS "Skipping example 7 because mfem was not found")
+    message(STATUS "Skipping parallel example because MPI was not found")
 endif ()
