### 2023-08-16 8415595
[examples] move traits examples at the end
