+++ examples/CMakeLists.txt
@@ -8,6 +8,13 @@ enable_testing()
 add_subdirectory(voxels)
 add_subdirectory(dem)
 
+find_package(MPI)
+if (MPI_FOUND)
+    add_subdirectory(parallel)
+else ()
+    message(STATUS "Skipping parallel example because MPI was not found")
+endif ()
+
 find_package(CGAL)
 if (CGAL_FOUND)
     add_subdirectory(cgal)
@@ -38,11 +45,4 @@ if (mfem_FOUND)
     add_subdirectory(mfem)
 else ()
     message(STATUS "Skipping mfem example because mfem was not found")
-endif ()
-
-find_package(MPI)
-if (MPI_FOUND)
-    add_subdirectory(parallel)
-else ()
-    message(STATUS "Skipping parallel example because MPI was not found")
 endif ()
