### 2023-08-16 13fe390
[examples] streamline folder / exe names
