+++ examples/CMakeLists.txt
@@ -17,7 +17,7 @@ endif ()
 
 find_package(CGAL)
 if (CGAL_FOUND)
-    add_subdirectory(cgal)
+    add_subdirectory(cgal_traits)
 else ()
     message(STATUS "Skipping cgal example because CGAL was not found")
 endif ()
@@ -25,24 +25,24 @@ endif ()
 find_package(dune-functions)
 if (dune-functions_FOUND)
     # unfortunately we have to repeat those here :/
-    # see CMakeLists.txt of example4 for more details
+    # see CMakeLists.txt of dune_traits for more details
     find_package(Threads)
     find_package(TBB)
-    add_subdirectory(dune)
+    add_subdirectory(dune_traits)
 else ()
     message(STATUS "Skipping dune example because Dune was not found")
 endif ()
 
 find_package(DOLFINX)
 if (DOLFINX_FOUND)
-    add_subdirectory(dolfinx)
+    add_subdirectory(dolfinx_traits)
 else ()
     message(STATUS "Skipping dolfinx example because dolfinx was not found")
 endif ()
 
 find_package(mfem CONFIG NAMES MFEM)
 if (mfem_FOUND)
-    add_subdirectory(mfem)
+    add_subdirectory(mfem_traits)
 else ()
     message(STATUS "Skipping mfem example because mfem was not found")
 endif ()
