### 2023-08-16 4c70ef2
[examples] add main readme example
