+++ examples/CMakeLists.txt
@@ -5,6 +5,8 @@ cmake_minimum_required(VERSION 3.18)
 project(gridformat_examples)
 
 enable_testing()
+
+add_subdirectory(analytical)
 add_subdirectory(voxels)
 add_subdirectory(dem)
 
