### 2023-08-16 aabb41a
[examples] add minimal example from paper
