### 2023-08-16 d4881eb
[examples][voxels] simplify code; refocus example
