+++ examples/parallel/README.md
@@ -7,9 +7,9 @@ This example illustrates how to read and write data from and to grid files in pa
 Essentially, what one has to do is to pass a communicator to the `Reader` and `Writer`
 classes such that they can communicate the required data. When passing a communicator
 to the reader, it will read parallel file formats such that each processor only reads
-its corresponding piece of the data. One can also read in the entire file including all
-pieces by passing in an instance of `GridFormat::NullCommunicator`, which signals that
-a parallel file should be read sequentially.
+its corresponding piece of the data. When constructing the reader without a communicator
+(or when explicitly passing `GridFormat::null_communicator`), it will read the entire file
+and merge all pieces.
 
 In the context of simulations, grid data structures typically require additional information
 on the processor boundaries, ghost cells, etc. Since this is usually strongly dependent on
