### 2023-08-16 cd41897
[paper] Add funding info TK
