### 2023-08-17 34461e9
[traits][dune] iterate over cells exposed by trait only
