### 2023-08-21 979419e
[test/examples/bin] adjust license header year
