### 2023-08-22 #5761 @danielskatz
@dglaeser - yes, the editor (whoever it is) will thank you :)

instructions are in the first comment in this issue:

> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
