### 2023-08-22 #5761 @danielskatz
@ IgorBaratta - would you be willing to review this submission for JOSS?
