### 2023-08-22 #5761 @lukeolson
> @lukeolson - would you be willing to review this submission for JOSS?

Yep, I can do that.
