### 2023-08-22 090bc8f
[vtk][xml] fix default active field setter
