### 2023-08-22 c597361
[vtk][xml] remove obsolete if check
