### 2023-08-23 #5778 @lukeolson
@dglaeser There is `LICENSE.txt` and `LICENSES/MIT.txt` (they are the same up to formatting) -- is there an easy way to have just one?  I guess the CC4 license is in there for a mesh example?   In any case... the repo meets the qualifications of needing a license in a text file... so this is a minor point.
