### 2023-08-24 #5778 @dglaeser
@lukeolson, the REUSE specification says that each used license has to be placed in a `LICENSES` folder (https://reuse.software/spec/#license-files), but then they didn't seem to be discovered by other services. That is why we decided to put the main license in the top folder, such that e.g. GitHub discovers and displays it directly in the `about` section. The actions check the repository for reuse compliance with the reuse tool, and the message of the commit that introduced `LICENSE.txt` says "The reuse specification allows this specifically to be compatible with other conventions."

> I guess the CC4 license is in there for a mesh example?

Yes, and also the images of the documenation are licensed with CC-BY-4
