### 2023-08-29 #5778 @danielskatz
👋 @IgorBaratta - when you get a chance, can you please generate your checklist by entering `@editorialbot generate my checklist` as a new comment here, then checking the COI and CoC criteria and checking off those items to make sure that everything is working right?
