### 2023-09-22 975dcf7
[traits][dune][functions] support normal grid views
