### 2023-09-22 2956ef8
[traits][dune][higher-order] speed-up
