+++ gridformat/traits/dune.hpp
@@ -384,6 +384,9 @@ namespace LagrangeDetail {
         static constexpr std::size_t reserved_size = 64;
 
      public:
+        using PointSet = DUNE::EquidistantPointSet<typename GridView::ctype, GridView::dimension>;
+        using Point = typename PointSet::LagrangePoint;
+
         explicit LocalPoints(unsigned int order)
         : _points(order)
         {}
@@ -399,7 +402,7 @@ namespace LagrangeDetail {
             return _points.size();
         }
 
-        const auto& at(std::size_t i) const {
+        const Point& at(std::size_t i) const {
             return _points[_sorted_indices.at(i)];
         }
 
@@ -428,12 +431,19 @@ namespace LagrangeDetail {
             });
         }
 
-        DUNE::EquidistantPointSet<typename GridView::ctype, GridView::dimension> _points;
+        PointSet _points;
         ReservedVector<unsigned int, reserved_size> _sorted_indices;
     };
 
-    class PointIndicesHelper {
+    class PointMapper {
      public:
+        template<typename GridView>
+        explicit PointMapper(const GridView& grid_view) {
+            _codim_to_global_indices.resize(GridView::dimension + 1);
+            for (int codim = 0; codim < GridView::dimension + 1; ++codim)
+                _codim_to_global_indices[codim].resize(grid_view.size(codim));
+        }
+
         struct Key {
             unsigned int codim;
             std::size_t global_index;
@@ -441,27 +451,34 @@ namespace LagrangeDetail {
         };
 
         bool contains(const Key& key) const {
-            if (!_codim_to_global_indices.contains(key.codim))
+            const auto& entity_dofs = _codim_to_global_indices[key.codim][key.global_index];
+            if (entity_dofs.size() <= key.sub_index)
                 return false;
-            if (!_codim_to_global_indices.at(key.codim).contains(key.global_index))
-                return false;
-            return _codim_to_global_indices.at(key.codim).at(key.global_index).contains(key.sub_index);
+            return entity_dofs[key.sub_index] != -1;
         }
 
         void insert(const Key& key, std::size_t index) {
-            _codim_to_global_indices[key.codim][key.global_index][key.sub_index] = index;
+            auto& entity_dofs = _codim_to_global_indices[key.codim][key.global_index];
+            if (entity_dofs.size() < key.sub_index + 1)
+                entity_dofs.resize(key.sub_index + 1, -1);
+            entity_dofs[key.sub_index] = static_cast<std::int64_t>(index);
         }
 
-        std::size_t get(const Key& key) {
-            return _codim_to_global_indices.at(key.codim).at(key.global_index).at(key.sub_index);
+        std::size_t get(const Key& key) const {
+            assert(_codim_to_global_indices[key.codim][key.global_index].size() > key.sub_index);
+            auto index = _codim_to_global_indices[key.codim][key.global_index][key.sub_index];
+            assert(index != -1);
+            return static_cast<std::size_t>(index);
+        }
+
+        void clear() {
+            _codim_to_global_indices.clear();
         }
 
      private:
-        std::unordered_map<
-            int, // codim
-            std::unordered_map<
-                std::size_t, // entity index
-                std::unordered_map<int, std::size_t>  // sub-entity index to global indices
+        std::vector<  // codim
+            std::vector<  // entity index
+                ReservedVector<std::int64_t, 20>  // sub-entity index to global indices
             >
         > _codim_to_global_indices;
     };
@@ -478,25 +495,37 @@ namespace LagrangeDetail {
  */
 template<typename GV>
[…]
