### 2023-09-24 aabcf78
[docs][traits] mention VTK point ordering
