+++ dune_8hpp.js
@@ -14,8 +14,8 @@ var dune_8hpp =
     [ "GridFormat::Traits::Origin< Dune::GridView< Traits > >", "structGridFormat_1_1Traits_1_1Origin_3_01Dune_1_1GridView_3_01Traits_01_4_01_4.html", null ],
     [ "GridFormat::Traits::Spacing< Dune::GridView< Traits > >", "structGridFormat_1_1Traits_1_1Spacing_3_01Dune_1_1GridView_3_01Traits_01_4_01_4.html", null ],
     [ "GridFormat::Traits::Ordinates< Dune::GridView< Traits > >", "structGridFormat_1_1Traits_1_1Ordinates_3_01Dune_1_1GridView_3_01Traits_01_4_01_4.html", null ],
-    [ "GridFormat::Dune::LagrangePolynomialGrid< GV >::Point", "structGridFormat_1_1Dune_1_1LagrangePolynomialGrid_1_1Point.html", null ],
-    [ "GridFormat::Dune::LagrangePolynomialGrid< GV >::Cell", "structGridFormat_1_1Dune_1_1LagrangePolynomialGrid_1_1Cell.html", null ],
+    [ "GridFormat::Dune::Traits::GridView< GV >", "structGridFormat_1_1Dune_1_1Traits_1_1GridView.html", null ],
+    [ "GridFormat::Dune::Traits::GridView< LagrangePolynomialGrid< GV > >", "structGridFormat_1_1Dune_1_1Traits_1_1GridView_3_01LagrangePolynomialGrid_3_01GV_01_4_01_4.html", null ],
     [ "GridFormat::Traits::Points< Dune::LagrangePolynomialGrid< GridView > >", "structGridFormat_1_1Traits_1_1Points_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_01_4_01_4.html", null ],
     [ "GridFormat::Traits::Cells< Dune::LagrangePolynomialGrid< GridView > >", "structGridFormat_1_1Traits_1_1Cells_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_01_4_01_4.html", null ],
     [ "GridFormat::Traits::NumberOfPoints< Dune::LagrangePolynomialGrid< GridView > >", "structGridFormat_1_1Traits_1_1NumberOfPoints_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_01_4_01_4.html", null ],
@@ -25,6 +25,7 @@ var dune_8hpp =
     [ "GridFormat::Traits::CellType< Dune::LagrangePolynomialGrid< GridView >, typename Dune::LagrangePolynomialGrid< GridView >::Cell >", "structGridFormat_1_1Traits_1_1CellType_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_01_4_00_062205e5ffe86918f1740aa68d141ed56.html", null ],
     [ "GridFormat::Traits::PointCoordinates< Dune::LagrangePolynomialGrid< GridView >, typename Dune::LagrangePolynomialGrid< GridView >::Point >", "structGridFormat_1_1Traits_1_1PointCoordinates_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_03f68cf5084bbc06f63c37b2704db23c7.html", null ],
     [ "GridFormat::Traits::PointId< Dune::LagrangePolynomialGrid< GridView >, typename Dune::LagrangePolynomialGrid< GridView >::Point >", "structGridFormat_1_1Traits_1_1PointId_3_01Dune_1_1LagrangePolynomialGrid_3_01GridView_01_4_00_01b42e2dd820d99e51ddd5157dd47d1c20.html", null ],
[…]
