+++ classGridFormat_1_1Dune_1_1LagrangePolynomialGrid.js
@@ -1,5 +1,0 @@
-var classGridFormat_1_1Dune_1_1LagrangePolynomialGrid =
-[
-    [ "Cell", "structGridFormat_1_1Dune_1_1LagrangePolynomialGrid_1_1Cell.html", null ],
-    [ "Point", "structGridFormat_1_1Dune_1_1LagrangePolynomialGrid_1_1Point.html", null ]
-];
