### 2023-09-24 25470c1
[traits][dune][functions] map id to running idx; map coordinates
