### 2023-09-24 0f0f9fe
[traits][dune][functions] skip evaluated points
