### 2023-09-29 397b173
[common][reservedvector] fix size of preallocated memory
