### 2023-09-29 bb1920e
[reserved_vector] enable move-construction
