### 2023-10-09 #5778 @IgorBaratta
Hi @danielskatz .

Review (almost) complete and everything looks good with the manuscript.
Just one small thing: there’s an unresolved issue regarding the documentation link, which is already flagged on GitHub ([Issue #162](https://github.com/dglaeser/gridformat/issues/162)). A fix or update on this would be sufficient before I check all the boxes.
