+++ docs/doxygen/markdown_filter.py
@@ -4,7 +4,7 @@
 import sys
 import string
 import subprocess
-from os.path import abspath, dirname, join, exists
+from os.path import abspath, dirname, join, exists, relpath
 
 
 THIS_DIR = dirname(__file__)
@@ -12,6 +12,12 @@ DOC_DIR = dirname(THIS_DIR)
 TOP_LEVEL_DIR = dirname(DOC_DIR)
 MAIN_README = abspath(join(TOP_LEVEL_DIR, "README.md"))
 assert exists(MAIN_README)
+
+
+def _get_commit_sha() -> str:
+    return subprocess.run(
+        "git rev-list --max-count=1 HEAD".split(" "), capture_output=True, text=True
+    ).stdout.strip("\n")
 
 
 def _filter_characters(text: str) -> str:
@@ -27,12 +33,30 @@ def _add_header_label(line: str) -> str:
     return f"{line} {{#{label}}}\n"
 
 
-def _process_line(line: str) -> str:
+def _handle_relative_urls(line: str, file_path: str) -> str:
+    pieces = line.split("<!-- DOXYGEN_MAKE_ABSOLUTE -->")
+    if len(pieces) == 1:
+        return line
+
+    def _make_absolute_url(p: str) -> str:
+        rel_path = p.split("](", maxsplit=1)[1].split(")", maxsplit=1)[0]
+        rel_path_from_top_level = relpath(join(dirname(file_path), rel_path.strip()), TOP_LEVEL_DIR)
+        return p.replace(
+            f"({rel_path})",
+            f"(https://github.com/dglaeser/gridformat/tree/{_get_commit_sha()}/{rel_path_from_top_level})"
+        )
+    return pieces[0] + "".join(_make_absolute_url(p) for p in pieces[1:])
+
+
+def _process_line(line: str, file_path: str) -> str:
     hint_key = "<!-- DOXYGEN_ONLY"
     if hint_key in line:
         return line.replace(hint_key, "").replace("-->", "").strip()
     else:
-        return _add_header_label(line)
+        return _add_header_label(
+            _handle_relative_urls(line, file_path)
+        )
+
 
 def _is_comment(line):
     return line.startswith("<!--")
@@ -59,6 +83,6 @@ if filepath == MAIN_README:
     content = f"# Introduction\n\n{content}"
 
 print("\n".join([
-    _process_line(line)
+    _process_line(line, filepath)
     for line in _remove_leading_comments_and_empty_lines(content.split("\n"))
 ]))
