### 2023-10-10 5027670
[actions][doxygen] flag repo as safe git dir
When building the docu, the markdown filter script tries to access
the git history to get the current commit sha. However, this fails
due to permission errors, probably because no git user is set in
the actions. Flagging the folder as safe solves the issue.
