+++ .github/workflows/pages.yml
@@ -20,7 +20,7 @@ jobs:
         uses: actions/checkout@v2
         with:
           submodules: true
-          
+
       - name: install-rsync
         run: apt-get update && apt-get install -y rsync
 
@@ -39,6 +39,7 @@ jobs:
 
       - name: build-doxygen-documentation
         run: |
+          git config --global --add safe.directory $PWD
           cd build && make doc_doxygen
           mkdir -p public && cp -r docs/doxygen/html/* public
 
