+++ paper/paper.bib
@@ -122,7 +122,7 @@ isbn={978-3-0348-8534-8}
 }
 
 @misc{mpi_web,
-  title = {{MPI} Forum},
+  title = {{MPI} {F}orum},
   howpublished = {Website: \url{https://www.mpi-forum.org/docs/}},
   year = {2023},
   note = {Accessed: 2023-06-22}
