### 2023-10-11 #5778 @danielskatz
@dglaeser - I've proofread the paper, and have a couple of minor changes to request in it, in https://github.com/dglaeser/gridformat/pull/222. Please merge this, or let me know what you disagree with, then we can proceed.
