### 2023-10-11 #5778 @dglaeser
@danielskatz, I merged https://github.com/dglaeser/gridformat/pull/222, thanks a lot for the suggestions!
