### 2023-10-11 0f31fd3
[test][reserved_vec] add unit test
