### 2023-10-11 c3e6279
[common][reserved_vec] construction from vec with different static size
