+++ gridformat/common/reserved_vector.hpp
@@ -13,6 +13,7 @@
 #include <vector>
 #include <utility>
 #include <memory_resource>
+#include <algorithm>
 #include <iterator>
 
 namespace GridFormat {
@@ -25,34 +26,44 @@ class ReservedVector {
     using value_type = T;
 
     ReservedVector() = default;
+    ReservedVector(std::size_t n, const T& r) : ReservedVector() { _elements.resize(n, r); }
+    ReservedVector(std::initializer_list<T>&& initList) : ReservedVector() { _copy_back(initList); }
 
-    ReservedVector(const ReservedVector& other) : ReservedVector() {
-        _elements.clear();
-        _elements.reserve(other.size());
-        std::copy(other._elements.begin(), other._elements.end(), std::back_inserter(_elements));
-    }
+    // We need both template and non-template variant, because if only the template is present, the
+    // compiler tries to use the default copy ctor (which is deleted). Same for move ctor & assignments.
+    template<std::size_t M>
+    ReservedVector(const ReservedVector<T, M>& other) : ReservedVector() { _copy_back(other); }
+    ReservedVector(const ReservedVector& other) : ReservedVector() { _copy_back(other); }
 
-    ReservedVector(std::size_t n, const T& r) : ReservedVector() {
-        _elements.resize(n, r);
-    }
-
-    ReservedVector(std::initializer_list<T>&& initList) : ReservedVector() {
-        _elements.resize(initList.size());
-        std::copy(initList.begin(), initList.end(), _elements.begin());
-    }
-
-    ReservedVector(ReservedVector&& other) : ReservedVector() {
-        _elements.clear();
-        _elements.reserve(other.size());
-        std::move(other._elements.begin(), other._elements.end(), std::back_inserter(_elements));
-    }
+    template<std::size_t M>
+    ReservedVector(ReservedVector<T, M>&& other) : ReservedVector() { _move_back(std::move(other)); }
+    ReservedVector(ReservedVector&& other) : ReservedVector() { _move_back(std::move(other)); }
 
     ReservedVector& operator=(const ReservedVector& other) {
         _elements = Vector{typename Vector::allocator_type{&_resource}};
-        _elements.reserve(other.size());
-        std::copy(other._elements.begin(), other._elements.end(), std::back_inserter(_elements));
+        _copy_back(other);
         return *this;
-    };
+    }
+
+    template<std::size_t M>
+    ReservedVector& operator=(const ReservedVector<T, M>& other) {
+        _elements = Vector{typename Vector::allocator_type{&_resource}};
+        _copy_back(other);
+        return *this;
+    }
+
+    ReservedVector& operator=(ReservedVector&& other) {
+        _elements = Vector{typename Vector::allocator_type{&_resource}};
+        _move_back(std::move(other));
+        return *this;
+    }
+
+    template<std::size_t M>
+    ReservedVector& operator=(ReservedVector<T, M>&& other) {
+        _elements = Vector{typename Vector::allocator_type{&_resource}};
+        _move_back(std::move(other));
+        return *this;
+    }
 
     void clear() { _elements.clear(); }
     std::size_t size() const { return _elements.size(); }
@@ -76,6 +87,18 @@ class ReservedVector {
     decltype(auto) at(std::size_t i) const { return _elements.at(i); }
 
  private:
+    template<std::ranges::sized_range R>
+    void _copy_back(const R& other) {
+        _elements.reserve(std::ranges::size(other));
+        std::ranges::copy(other, std::back_inserter(_elements));
+    }
+
+    template<std::ranges::sized_range R> requires(!std::is_lvalue_reference_v<R>)
+    void _move_back(R&& other) {
[…]
