### 2023-10-11 55a3857
[util] add script to check version consistency
