+++ .github/workflows/main.yml
@@ -13,6 +13,7 @@ on:
   schedule:
     - cron: '30 22 * * 0,3'
   workflow_dispatch:
+  workflow_call:
 
 env:
   DEFAULT_COMPILER_PKG: g++-12
