+++ .github/workflows/release.yml
@@ -16,7 +16,7 @@ jobs:
     - name: check-version-consistency
       run: |
         git config --global --add safe.directory $PWD
-        python3 util/check_version_consistency.py
+        python3 util/update_versions.py --check-only -v ${{ github.ref_name }}
 
   test:
     needs: [version-sanity-check]
@@ -35,7 +35,7 @@ jobs:
     - uses: actions/checkout@v3
     - name: create-description-body
       run: |
-        python3 util/write_release_description.py -o ${DESCRIPTION_FILE} -v github.ref_name
+        python3 util/write_release_description.py -o ${DESCRIPTION_FILE} -v ${{ github.ref_name }}
     - name: build-doxygen-documentation
       run: |
         git config --global --add safe.directory $PWD
