### 2023-10-11 b820869
[workflows] test on release branches as well
