+++ .github/workflows/main.yml
@@ -5,9 +5,13 @@ name: test-suite
 
 on:
   push:
-    branches: [ main ]
+    branches:
+    - main
+    - 'releases/**'
   pull_request:
-    branches: [ main ]
+    branches:
+    - main
+    - 'releases/**'
     paths-ignore:
       - 'docker/**'
   schedule:
