### 2023-10-11 07dbdb6
[README] add section on release procedure
