### 2023-10-11 4171122
[README] mention version bump on main after release
