### 2023-10-11 #5778 @dglaeser
@danielskatz, I am currently setting up the Zenodo-GitHub integration such that releases on github are automatically also put up on Zenodo. However, the default is that the repository name is chosen as "title", while the "description" would match the title of the paper (this is taken from the about section on github). Authors are automatically transferred from the `CITATION.cff` file in the repository, which in our case means that one of the authors of the paper does not appear.

If this is incompatible with your requirements, I think I can manually modify the metadata of the automatically created dataset (also the title), or just do a manual release for this first one.
