### 2023-10-11 #5778 @danielskatz
JOSS authors typically edit the metadata (title and authors) manually after the zenodo archive is created
