+++ util/update_versions.py
@@ -4,6 +4,7 @@
 import sys
 import subprocess
 import argparse
+import datetime
 
 from os.path import join, abspath, dirname
 
@@ -61,23 +62,39 @@ def _set_cmake_version(new_version: str) -> None:
         cmake_file.write(f"{before}{new_version}{after}")
 
 
-def _split_cff_version() -> tuple[str, str, str]:
+def _split_cff_at(key: str) -> tuple[str, str, str]:
     with open(CFF_FILE_PATH) as cff_file:
         content = cff_file.read()
-        before, version_and_rest = _split_after("\nversion: ", content)
-        version, rest = version_and_rest.split("\n", maxsplit=1)
-        return before, version, f"\n{rest}"
+        before, value_and_rest = _split_after(key, content)
+        value, rest = value_and_rest.split("\n", maxsplit=1)
+        return before, value, f"\n{rest}"
 
 
 def _get_cff_version() -> str:
-    return _split_cff_version()[1]
+    return _split_cff_at("\nversion: ")[1]
 
 
 def _set_cff_version(new_version: str) -> None:
     assert _is_release_version_tag(f"v{new_version}")
-    before, _, after = _split_cff_version()
+    before, _, after = _split_cff_at("\nversion: ")
     with open(CFF_FILE_PATH, "w") as cff_file:
         cff_file.write(f"{before}{new_version}{after}")
+
+
+def _get_cff_release_date() -> datetime.date:
+    return datetime.date.fromisoformat(_split_cff_at("\ndate-released: ")[1].strip("'"))
+
+
+def _set_cff_release_date(date: datetime.date) -> None:
+    before, _, after = _split_cff_at("\ndate-released: ")
+    with open(CFF_FILE_PATH, "w") as cff_file:
+        cff_file.write(f"{before}'{date}'{after}")
+
+
+def _get_tag_date(tag: str) -> datetime.date:
+    tag_date = _run_and_capture(f"git log {tag}").split("\nDate: ", maxsplit=1)[1].split("\n", maxsplit=1)[0]
+    tag_date = " ".join(tag_date.split()[:-1])  # remove timezone
+    return datetime.datetime.strptime(tag_date, "%a %b %d %H:%M:%S %Y").date()
 
 
 if __name__ == "__main__":
@@ -100,7 +117,14 @@ if __name__ == "__main__":
     version = tag[1:]
     if args["check_only"]:
         if not args["skip_tag"]:
+            branch = _run_and_capture("git rev-parse --abbrev-ref HEAD")
             _run_and_capture(f"git checkout {tag}")
+            cff_date = _get_cff_release_date()
+            tag_date = _get_tag_date(tag)
+            _run_and_capture(f"git switch {branch}")
+            if cff_date != tag_date:
+                sys.stderr.write("Tag date does not match cff release date")
+                sys.exit(1)
         cff_version = _get_cff_version()
         cmake_version = _get_cmake_version()
         if cff_version != version:
@@ -110,7 +134,11 @@ if __name__ == "__main__":
             sys.stderr.write(f"Version specified in cmake does not match the current tag version ({cmake_version})\n")
             sys.exit(1)
     else:
-        if not args["skip_tag"]:
-            _run_and_capture(["git", "tag", f"{tag}", "HEAD", "-a", "-m", f"Tag version {version}"])
         _set_cmake_version(version)
         _set_cff_version(version)
+        _set_cff_release_date(datetime.datetime.now().date())
+        if not args["skip_tag"]:
+            subprocess.run([
+                "git", "commit", "-m", f"bump version to {tag}", CFF_FILE_PATH, CMAKE_LISTS_PATH
+            ], check=True)
+            _run_and_capture(["git", "tag", f"{tag}", "HEAD", "-a", "-m", f"Tag version {version}"])
