+++ README.md
@@ -219,17 +219,17 @@ If your Python environment does not have `VTK`, this step is skipped. Note that 
 
 ### Creating a release
 
-To create a release, you may use the utility script `util/update_versions.py`, which creates a git tag and adjusts the versions
-specified in the cmake setup and the `CITATION.cff` file. For each release we maintain a separate branch for bugfixes and patch
-releases. As an example, to create a release version `1.2.3`, you may type the following into the console (assumes a clean repository):
+To create a release, you may use the utility script `util/update_versions.py`, which creates a git tag and adjusts the versions and
+release dates specified in the cmake setup and the `CITATION.cff` file. For each release we maintain a separate branch for bugfixes
+and patch releases. As an example, to create a release version `1.2.3`, you may type the following into the console (assumes a clean
+repository):
 
 ```bash
 git switch --create releases/1.2.3
 # ... maybe continue development ...
-python3 util/update_versions.py -v 1.2.3
-git commit -m "bump version to v1.2.3" .
+python3 util/update_versions.py -v 1.2.3 # modifies versions&dates and creates a commit + tag
 git push origin releases/1.2.3
-git push origin v1.2.3  # this tag was created by the script
+git push origin v1.2.3
 ```
 
 Afterwards, a release workflow will be triggered. If this runs through successfully, a release has been created. If not, the new tag
@@ -239,7 +239,7 @@ should be increased (with a suffix `-git`). Following the above example, you may
 ```bash
 git switch main
 git switch --create feature/bump-version
-python3 util/update_versions.py -v 1.2.4-git --skip-tag
+python3 util/update_versions.py -v 1.2.4-git --skip-tag # only modifies versions, no commit or tag
 git commit -m "bump version to v1.2.3" .
 git push origin feature/bump-version
 ```
