### 2023-10-12 a9e3be5
[util][version] switch back only if valid branch
