### 2023-10-12 537e306
[util][release-description] support arg in the form vX.X.X
