+++ .github/workflows/main.yml
@@ -5,9 +5,7 @@ name: test-suite
 
 on:
   push:
-    branches:
-    - main
-    - 'releases/**'
+    branches: [main]
   pull_request:
     branches:
     - main
