### 2023-10-12 d81d720
[actions][release] configure before building docs
