### 2023-10-12 732c0f2
[actions][release] add extension to artifact
