+++ CITATION.cff
@@ -29,5 +29,5 @@ keywords:
   - grid file I/O
   - mesh file I/O
 license: MIT
-version: 0.1.0
-date-released: '2023-10-11'
+version: 0.1.1
+date-released: '2023-10-12'
