+++ navtreedata.js
@@ -45,7 +45,8 @@ var NAVTREE =
     ] ],
     [ "Files", "files.html", [
       [ "File List", "files.html", "files_dup" ]
-    ] ]
+    ] ],
+    [ "Documentation for other versions", "^https://github.com/dglaeser/gridformat/releases", null ]
   ] ]
 ];
 
