### 2023-10-13 5bb83f8
[actions][release] use context variables
