+++ CITATION.cff
@@ -30,4 +30,4 @@ keywords:
   - mesh file I/O
 license: MIT
 version: 0.1.0
-date-released: '2023-10-11'
+date-released: '2023-10-13'
