+++ CITATION.cff
@@ -29,5 +29,5 @@ keywords:
   - grid file I/O
   - mesh file I/O
 license: MIT
-version: 0.1.1
-date-released: '2023-10-12'
+version: 0.1.2
+date-released: '2023-10-13'
