+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 <!--SPDX-FileCopyrightText: 2023 Dennis Gläser <dennis.glaeser@iws.uni-stuttgart.de>-->
 <!--SPDX-License-Identifier: MIT-->
 
+
+# `GridFormat` 0.1.1
+
+## Fixes
+
+- Fixed the artifact path in the release workflow.
+
 # `GridFormat` 0.1.0
 
 Our very first release 🎉
