+++ CITATION.cff
@@ -30,4 +30,4 @@ keywords:
   - mesh file I/O
 license: MIT
 version: 0.1.1
-date-released: '2023-10-12'
+date-released: '2023-10-13'
