### 2023-10-13 c1ed2ad
[util] fix description extraction from changelog
