+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 <!--SPDX-FileCopyrightText: 2023 Dennis Gläser <dennis.glaeser@iws.uni-stuttgart.de>-->
 <!--SPDX-License-Identifier: MIT-->
 
+# `GridFormat` 0.1.2
+
+## Fixes
+
+- Fixed description extraction from changelog for releases.
 
 # `GridFormat` 0.1.1
 
