+++ CITATION.cff
@@ -30,4 +30,4 @@ keywords:
   - mesh file I/O
 license: MIT
 version: 0.1.2
-date-released: '2023-10-13'
+date-released: '2023-10-16'
