+++ CITATION.cff
@@ -29,5 +29,5 @@ keywords:
   - grid file I/O
   - mesh file I/O
 license: MIT
-version: 0.1.2
-date-released: '2023-10-13'
+version: 0.1.3
+date-released: '2023-10-16'
