+++ CITATION.cff
@@ -31,3 +31,7 @@ keywords:
 license: MIT
 version: 0.1.3
 date-released: '2023-10-16'
+identifiers:
+  - description: Collection of archived snapshots of all versions
+    type: doi
+    value: "10.5281/zenodo.10008061"
