### 2023-10-16 #5778 @dglaeser
@danielskatz, the Zenodo integration is set up now. There now is:

- a zenodo record at https://zenodo.org/records/10008062, DOI: 10.5281/zenodo.10008062, which references the
- git tag `v0.1.2` (see https://github.com/dglaeser/gridformat/tree/v0.1.2)

Let me know if there is anything that we should add. Thanks!
